got Launcher looping through main function now on to working on getting microserverses threadded
