+++ src/prosessing/Config.ini
@@ -38,10 +38,10 @@ Stream_port= 5000
 
 camera_fps= 15
 
-#loggging Config
+#loggging Config dont add file extendtion
 [LOGGING]
-filename = OpencvServer.log
-launcher = Launcher.log
+filename = OpencvServer
+launcher = Launcher
 # handles setting zmq port and ip
 [ZMQ]
 ip = 127.0.0.1
